Create Best_divisor.java
Kristen loves playing with and comparing numbers. She thinks that if she takes two different positive numbers, the one whose digits sum to a larger number is better than the other. If the sum of digits is equal for both numbers, then she thinks the smaller number is better. For example, Kristen thinks that 13  is better than 31 and that 12 is better than 11 .
#HackerRank Problem Best divisor