Create Special_string_again.java
Naive solution to Solve special string again problem in O(n*n*n).
A string is said to be a special string if either of two conditions is met:

All of the characters are the same, e.g. aaa.
All characters except the middle one are the same, e.g. aadaa.
A special substring is any substring of a string which meets one of those criteria. 

#HackerRank